requirements.txt
oracledb>=1.4.0
boto3>=1.28.0
python-dotenv>=1.0.0
